Replace materialized view with movies task
Currently, we use a materialized view to fetch the movie rankings
quickly. This will be a problem when adding more ratings, as the
whole view will need to be replaced. This removes the materialized view
and instead extends the movie database with the aggregated values
of ratings via a rake tasks.

The rake task is probably only run after the database is populated.
After that the database can be updated simply as part of the post
request that adds the rating.